+++ README.md
@@ -1,0 +1,23 @@
+# Flappy Birds - A fun game build using pygame
+
+# Installation
+On Linux –
+Install these pygame dependencies
+```shell
+$ sudo apt-get install python3-dev python3-numpy libsdl-dev libsdl-image1.2-dev libsdl-mixer1.2-dev libsdl-ttf2.0-dev libsmpeg-dev libportmidi-dev libavformat-dev libswscale-dev libjpeg-dev libfreetype6-dev
+```
+
+Then create a conda env
+
+```shell
+$ git clone https://github.com/MajorOS-Geeks/Flappy_Birds.git
+$ conda create --name flappy_game --file conda_req.txt
+$ cd Flappy_Birds
+$ conda activate flappy_game
+$ python main.py
+```
+ENJOY PLAYING THE GAME
+
+# The game is actually developed by [Anuj Vijjan](https://github.com/AnujVijjan)
+This organization was created for a small college project, unfortunately [Anuj Vijjan](https://github.com/AnujVijjan) due to his busy schedule hasn’t been able to join. In fact, he is busy developing new apps for the organization. So we are managing the repos here and making changes likewise.
+# All Contributions are accepted.
