+++ conda_reqs.txt
@@ -1,0 +1,22 @@
+# This file may be used to create an environment using:
+# $ conda create --name <env> --file <this file>
+# platform: linux-64
+_libgcc_mutex=0.1=main
+ca-certificates=2020.1.1=0
+certifi=2019.11.28=py38_1
+libedit=3.1.20181209=hc058e9b_0
+libffi=3.2.1=hd88cf55_4
+libgcc-ng=9.1.0=hdf63c60_0
+libstdcxx-ng=9.1.0=hdf63c60_0
+ncurses=6.2=he6710b0_0
+openssl=1.1.1f=h7b6447c_0
+pip=20.0.2=py38_1
+pygame=1.9.6=pypi_0
+python=3.8.0=h0371630_2
+readline=7.0=h7b6447c_5
+setuptools=46.1.3=py38_0
+sqlite=3.31.1=h7b6447c_0
+tk=8.6.8=hbc83047_0
+wheel=0.34.2=py38_0
+xz=5.2.4=h14c3975_4
+zlib=1.2.11=h7b6447c_3
