+++ gallery/sprites/base.png
