+++ gallery/sprites/pipe.png
